Small file reorg (only renames/moves) (#271)
* reorg utils

* push missing files

* style

* fix
