+++ src/lighteval/logging/hierarchical_logger.py
@@ -26,7 +26,7 @@ from datetime import timedelta
 from logging import Logger
 from typing import Any, Callable
 
-from lighteval.utils import is_accelerate_available, is_nanotron_available
+from lighteval.utils.imports import is_accelerate_available, is_nanotron_available
 
 
 if is_nanotron_available():
