+++ src/lighteval/metrics/__init__.py
@@ -25,7 +25,7 @@ import re
 from lighteval.metrics.metrics import Metric, MetricCategory
 from lighteval.models.model_output import ModelResponse
 from lighteval.tasks.requests import Doc
-from lighteval.utils import as_list
+from lighteval.utils.utils import as_list
 
 
 def apply_target_perplexity_metric(results: list[ModelResponse], formatted_doc: Doc, metrics: list[Metric]):
