+++ src/lighteval/metrics/metrics_corpus.py
@@ -35,7 +35,7 @@ from lighteval.metrics.sample_preparator import (
     LogprobCorpusMetricInput,
     PerplexityCorpusMetricInput,
 )
-from lighteval.utils import as_list
+from lighteval.utils.utils import as_list
 
 
 # General aggregations
