+++ src/lighteval/models/adapter_model.py
@@ -29,7 +29,8 @@ from lighteval.logging.hierarchical_logger import hlog
 from lighteval.models.base_model import BaseModel
 from lighteval.models.model_config import AdapterModelConfig
 from lighteval.models.utils import _get_dtype
-from lighteval.utils import EnvConfig, is_peft_available
+from lighteval.utils.imports import is_peft_available
+from lighteval.utils.utils import EnvConfig
 
 
 if is_peft_available():
