+++ src/lighteval/models/delta_model.py
@@ -30,7 +30,7 @@ from lighteval.logging.hierarchical_logger import hlog
 from lighteval.models.base_model import BaseModel
 from lighteval.models.model_config import DeltaModelConfig
 from lighteval.models.utils import _get_dtype
-from lighteval.utils import EnvConfig
+from lighteval.utils.utils import EnvConfig
 
 
 class DeltaModel(BaseModel):
