+++ src/lighteval/models/model_loader.py
@@ -39,7 +39,7 @@ from lighteval.models.model_config import (
     TGIModelConfig,
 )
 from lighteval.models.tgi_model import ModelClient
-from lighteval.utils import NO_TGI_ERROR_MSG, is_tgi_available
+from lighteval.utils.imports import NO_TGI_ERROR_MSG, is_tgi_available
 
 
 def load_model(  # noqa: C901
