+++ src/lighteval/tasks/registry.py
@@ -35,7 +35,7 @@ import lighteval.tasks.default_tasks as default_tasks
 from lighteval.logging.hierarchical_logger import hlog, hlog_warn
 from lighteval.tasks.extended import AVAILABLE_EXTENDED_TASKS_MODULES
 from lighteval.tasks.lighteval_task import LightevalTask, LightevalTaskConfig
-from lighteval.utils import CANNOT_USE_EXTENDED_TASKS_MSG, can_load_extended_tasks
+from lighteval.utils.imports import CANNOT_USE_EXTENDED_TASKS_MSG, can_load_extended_tasks
 
 
 # Helm, Bigbench, Harness are implementations following an evaluation suite setup
